test crc.c x86_64 compile fixes
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@415 048eaad1-c872-4173-8c30-523bb29f53ec

+++ crc.h
@@ -25,7 +25,7 @@
 
 #if defined(CRC_CCITT)
 
-typedef uint8_t  crc;
+typedef unsigned short  crc;
 
 #define CRC_NAME			"CRC-CCITT"
 #define POLYNOMIAL			0x1021
